Failure message now prints the content of the web page if the failure is a web "server error" this special case was necessary because the default str method of the web error object does not print the content of the page.
cashOut special case handling fix : if cashOut all money from the safe. In this case the
currency server is expected to return two notes : one of the total amount and the other with
zero. The zero value note is discarded and the code gracefully behaves when there is no bank
note going back to the safe (i.e. the user has a note and there's nothing left in the safe).