PokerService constructor accepts a string instead of a Config object. The string is converted into a Config object. For testing purposes.
cancelTimers functions that removes all timers with a name containing the 
string argument, for use by tournament break timers when shutting down the
server. 

tourneyNewState method handles tournament break specific states by calling
dedicated methods for each case.

new tourneyBreakCheck method is a timer method called while the tournament is on break
to inform the user and exit from the tournament break state.

new tourneyDeal method that is called when a tournament switches to
running state because in rare cases it may not receive any event that
triggers the PokerTable autodeal function (such a case shows while
running the tests under a controlled clock)

new tourneyBreakWait method that broadcast informative messages to the players involved
in the tournament, depending on their involvement in a hand.

new tourneyBreakResume method that broadcasts a messages to all players when the tournament
exits from the break state to go back to the running state.