switch from get_message() => search_output in two places.
In PokerServiceTestCase.test15_runTourney() and
TourneyManagerTestCase.test01_no_rank(), switch from an exact check of
get_messages() to a series of search_output()s.  It is possible that extra
verbosity in output from modules called by PokerService.tourneyRegister()
(such as the database module) can cause get_messages() to have extra
messages in it that should be ignored by test and not cause failure.