Properly implemented min_players feature for PokerService.listTables() and corrected/augmented corresponding test.  This work is towards sr#2364.
I realized that user2table shouldn't be used for finding minimum players,
but rather the players field in pokertables, since that is updated by the
underlying module, and saves an unnecessary join in the SQL statement and
more difficult building of the SQL string.  That required numerous changes
to the SQL statements in the test.  There were also some problems with the
test, mostly related to what happens when players is 0.

A few other tests also needed an extra \t field added, now that an
additional one is properly supported.