Slight refactor of handlePacketLogic() to use sub-methods for certain packet operations.
The "perform" methods are designed so that the a minimal amount of code
related to receiving a packet that appears in the handlePacketLogic()
giant "if" statement.  The primary motive for this is for things like
PacketTablePicker(), that need to perform operations *as if* the client
has sent additional packets.  The desire is to keep completely parity
between what the individual packets do by themselves, and what
"super-packets" like PacketTablePicker() do.  A secondary benefit is that
it makes that giant "if" statement in handlePacketLogic() above a bit
smaller.

This is work toward competition of feature request in sr#2364.