Basic implementation of a test that exhibits bug#14139.
This new test file launches a bot server, and waits for the bot to join
the table.  Then, it has a real user sit.  There is a specific failure
that should occur until bug#14139 is closed.

Some changes were also needed in test-pokeravatar.py.in, since we
obviously wanted to reuse some of the helper methods from those base
classes.  We have confirmed that these changes don't cause any of the
actual tests in test-pokeravatar.py.in still pass despite these changes.