Test to exhibit bug#14139.
This new test in test-bug-14139.py.in exhibits bug#14139 by adding bots
and user and showing that a hand does not start.  If the bug is closed,
this test should pass as written.

Note that some changes were needed in test-pokeravatar.py.in in some of
the helper functions.  This doesn't impact the functioning of those
test.  Much of the changes move some methods to base classes; the rest
make minor changes to helper functions to handle the existence of the
bots at the tables and slightly different uses of the helper functions.

I've also begun what will (hopefully) become a common tradition of adding
test files to XFAIL_TESTS when they are designed for bugs that are still
open.  As bugs get closed, the test can be removed form XFAIL_TESTS.

Finally, a few files that should have been in .gitignore are added.
This will be meaningless currently to anyone but those using git-svn.